requirements.txt
# Video Processing
moviepy>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
imageio-ffmpeg>=0.4.9

# TTS & Transcription
openai>=1.0.0
httpx>=0.24.0
aiohttp>=3.8.0

# Instagram Downloader
playwright>=1.40.0
aiofiles>=23.0.0
aiohttp-socks>=0.8.0
aiodns>=3.0.0

# Utils
python-dotenv>=1.0.0
